Using reusable component for both editing and adding student
