+++ FrontEnd/src/Components/Profile/StudentSection.jsx
@@ -5,8 +5,7 @@ import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
 import AddIcon from '@mui/icons-material/Add';
 import StudentItem from './StudentItem';
-import StudentProfile from '../../Pages/ProfilePage/StudentProfileFul';
-import EditStudentProfile from '../../Pages/ProfilePage/EditStudentProfile';
+import StudentProfile from '../../Pages/ProfilePage/StudentProfile';
 import DeleteDialog from './DeleteDialog';
 import { deleteStudent, findStudents } from '../../api/students';
 import { defaultStudentInfo } from '../../utils/constants';
@@ -33,7 +32,6 @@ function StudentSection({
     handleOpenDeleteDialogState();
     deleteStudent({ _id: currentStudent._id })
       .then((data) => {
-        console.log(data);
         if (data.status === 400) {
           setAlertMessage('Se produjo un error al eliminar al estudiante.');
           setErrorOpen(true);
@@ -46,11 +44,18 @@ function StudentSection({
         findStudents({ idUser: user.sub })
           .then((response) => response.json())
           .then((data) => {
-            console.log(data);
             setStudents(data);
           });
       });
   };
+
+  const handleCloseStudentModal = () => {
+    if (addStudent) setAddStudent(!addStudent);
+    else {
+      setOpenEditModal(!openEditModal);
+      setIsEditing(false);
+    }
+  }
 
   const canAddStudents = user?.p99roles.length === 0;
 
@@ -59,7 +64,7 @@ function StudentSection({
       <Box
         sx={{
           pt: 2,
-          display: canAddStudents? 'flex' : 'none',
+          display: canAddStudents ? 'flex' : 'none',
           justifyContent: 'space-between',
           mt: 2,
         }}
@@ -69,7 +74,7 @@ function StudentSection({
       <Box>
         {students !== null
         && students.length === 0
-        && canAddStudents? (
+        && canAddStudents ? (
           <Box
             sx={{
               fontFamily: 'default',
@@ -99,7 +104,7 @@ function StudentSection({
           sx={{
             display: {
               xs: 'none',
-              md: canAddStudents? 'flex' : 'none',
+              md: canAddStudents ? 'flex' : 'none',
               justifyContent: 'flex-end',
             },
           }}
@@ -123,46 +128,23 @@ function StudentSection({
       </Box>
 
       <Modal
-        open={addStudent}
-        onClose={() => setAddStudent(!addStudent)}
+        open={addStudent || openEditModal}
+        onClose={handleCloseStudentModal}
         sx={{ overflow: 'scroll' }}
       >
-        <>
-          <StudentProfile
-            studentInfo={defaultStudentInfo}
-            setOpenStudentProfile={setAddStudent}
-            openStudentProfile={addStudent}
-            userID={user.sub}
-            setStudents={setStudents}
-            isEditing={true}
-            setIsEditing={() => {}}
-            setSuccessOpen={setSuccessOpen}
-            setErrorOpen={setErrorOpen}
-            setAlertMessage={setAlertMessage}
-            setInfoOpen={setInfoOpen}
-          />
-        </>
-      </Modal>
-
-      <Modal
-        open={openEditModal}
-        onClose={() => {setOpenEditModal(!openEditModal); setIsEditing(false);}}
-        sx={{ overflow: 'scroll' }}
-      >
-        <>
-          <StudentProfile
-            openStudentProfile={openEditModal}
-            setOpenStudentProfile={setOpenEditModal}
-            studentInfo={currentStudent}
-            setIsEditing={setIsEditing}
-            isEditing={isEditing}
-            setStudents={setStudents}
-            setSuccessOpen={setSuccessOpen}
-            setErrorOpen={setErrorOpen}
-            setAlertMessage={setAlertMessage}
-            setInfoOpen={setInfoOpen}
-          />
-        </>
+        <StudentProfile
+          studentInfo={(addStudent ? defaultStudentInfo : currentStudent)}
+          openStudentProfile={(addStudent || openEditModal)}
+          setOpenStudentProfile={(addStudent ? setAddStudent : setOpenEditModal)}
+          userID={(addStudent ? user.sub : undefined)}
+          setStudents={setStudents}
+          isEditing={(addStudent ? true : isEditing)}
+          setIsEditing={(addStudent ? () => {} : setIsEditing)}
+          setSuccessOpen={setSuccessOpen}
+          setErrorOpen={setErrorOpen}
+          setAlertMessage={setAlertMessage}
+          setInfoOpen={setInfoOpen}
+        />
       </Modal>
 
       <DeleteDialog
